KlphaOriginal/pyproject.toml
[project]
name = "klpha"
version = "0.1.0"
description = "美股横截面因子量化交易平台"
requires-python = ">=3.9"
dependencies = [
    "pandas>=1.5.0",
    "numpy>=1.20.0",
    "yfinance>=0.2.0",
    "pyyaml>=6.0",
    "openpyxl>=3.0.0",  # 用于 pandas read_html
    "numba>=0.56.0",  # 用于加速计算
    "matplotlib>=3.5.0",  # 用于绘图（可选）
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

